Extract AutoAcceptClientStore to its own file
First step of the Javalin decoupling work tracked in #72's follow-up
plan. AutoAcceptClientStore had no HTTP dependencies but was nested
inside FakeIdProvider; lifting it out now so the upcoming core/server
split can move it unchanged.

No behavioural change — all 25 tests still pass.
